Keep field arithmetic inside the range it is allowed to use
Fuzzing against a build with the undefined behaviour sanitizer found three ways to reach
arithmetic the calendars cannot represent: adding to a field worked out the new value before
checking it, negating the amount for an era that counts backwards did not fit at the bottom of
the range, and a floor division took its remainder from a product wider than the numerator it
came from. All three now use the checked arithmetic the field resolution already had.

A calendar that cannot represent an instant reported it as an internal error, which takes the
whole database down rather than failing the query that asked for it. It is a conversion error
now.

Fixing the first of those exposed a fourth: a failure was cleared by the next call rather than
kept, so date_add discarded it and adding two billion months returned a date in the past. An
operation now runs from SetTime to the read at the end of it.

+++ extension/icu/datetime/include/grego.hpp
@@ -48,7 +48,9 @@ struct FloorDiv {
 	}
 	static int32_t Divide(int32_t numerator, int32_t denominator, int32_t &remainder) {
 		const auto quotient = Divide(numerator, denominator);
-		remainder = numerator - (quotient * denominator);
+		// the remainder is always smaller than the denominator, but the product it is taken
+		// from is not, so it is formed wide enough to hold the extremes of the numerator
+		remainder = int32_t(int64_t(numerator) - int64_t(quotient) * denominator);
 		return quotient;
 	}
 	//! Divides a millisecond count by a whole number of milliseconds, keeping the remainder exact
